Remove Gemfile.lock
Given that this project is a gem that will be included into other
projects, it's best to leave out Gemfile.lock and hence make us always
test with the latest version of the dependencies.

This will also disable a false alarm from GitHub about vulnerability, as
this gem does not really use those dependencies.